requirements.txt
tqdm
Flask
Werkzeug
numpy
torch
torchvision
opencv-python
matplotlib
Pillow
vietocr
paddleocr
requests
craft-text-detector